refactor: 게임 결과를 생성하는 로직을 CardGameResult로 이동
